Rename test files and update references
