+++ src/test/directives/class-map_test.ts
@@ -15,7 +15,7 @@
 /// <reference path="../../../node_modules/@types/mocha/index.d.ts" />
 /// <reference path="../../../node_modules/@types/chai/index.d.ts" />
 
-import {ClassInfo, classMap} from '../../directives/classMap.js';
+import {ClassInfo, classMap} from '../../directives/class-map.js';
 import {render} from '../../lib/render.js';
 import {html} from '../../lit-html.js';
