+++ test/index.html
@@ -22,8 +22,8 @@
     <script type="module" src="./directives/until_test.js"></script>
     <script type="module" src="./directives/unsafe-html_test.js"></script>
     <script type="module" src="./directives/when_test.js"></script>
-    <script type="module" src="./directives/classMap_test.js"></script>
-    <script type="module" src="./directives/styleMap_test.js"></script>
+    <script type="module" src="./directives/class-map_test.js"></script>
+    <script type="module" src="./directives/style-map_test.js"></script>
     <script>
       WCT.loadSuites([
         'shady.html',
